@@ -36,38 +36,40 @@ dependencies:
   cupertino_icons: ^1.0.2
   riverpod: ^2.3.5
   riverpod_annotation: ^2.1.0
-  flutter_hooks: ^0.18.6
+  flutter_hooks: ^0.20.0
   hooks_riverpod: ^2.3.5
   freezed_annotation: ^2.2.0
   json_annotation: ^4.8.0
   wifi_scan: ^0.4.0+1
-  flutter_form_builder: ^7.3.1
-  form_builder_validators: ^8.1.1
+  flutter_form_builder: ^9.1.0
+  form_builder_validators: ^9.0.0
   flutter_blue_plus: ^1.1.3
-  amplify_flutter: ^1.0.1
-  amplify_datastore: ^1.0.0-supports-only-mobile
-  amplify_auth_cognito: ^1.0.1
-  amplify_authenticator: ^1.0.1 
-  amplify_api: ^1.0.1
+  amplify_flutter: ^1.3.3
+  amplify_datastore: ^1.3.4
+  amplify_auth_cognito: ^1.3.3
+  amplify_authenticator: ^1.3.2
+  amplify_api: ^1.3.2
   percent_indicator: ^4.2.3
   # amplify_core: ^1.0.1
-  # flutter_zxing: ^1.1.2
-  mqtt_client: ^9.8.1
-  amazon_cognito_identity_dart_2: ^3.3.0
+  flutter_zxing: ^1.3.2
+  mqtt_client: ^10.0.0
+  amazon_cognito_identity_dart_2: ^3.5.0
   go_router: ^10.1.0
   # aws_common: ^0.1.1
   # aws_signature_v4: ^0.2.0
   sigv4: ^5.0.0
   tuple: ^2.0.1
+  flutter_layout_grid: ^2.0.4
+  flutter_screenutil: ^5.9.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
+  device_preview: ^1.1.0
   riverpod_generator: ^2.2.0
   build_runner: ^2.3.3
-  riverpod_lint: ^1.3.0
-  custom_lint: ^0.3.4
+  riverpod_lint: ^2.0.1
+  custom_lint: ^0.5.2
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
