@@ -1,303 +1,303 @@
-// Copyright 2017, Paul DeMarco.
-// All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
+// // Copyright 2017, Paul DeMarco.
+// // All rights reserved. Use of this source code is governed by a
+// // BSD-style license that can be found in the LICENSE file.
 
-import 'package:flutter/material.dart';
-import 'package:flutter_blue_plus/flutter_blue_plus.dart';
+// import 'package:flutter/material.dart';
+// import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 
-class ScanResultTile extends StatelessWidget {
-  const ScanResultTile({Key? key, required this.result, this.onTap})
-      : super(key: key);
+// class ScanResultTile extends StatelessWidget {
+//   const ScanResultTile({Key? key, required this.result, this.onTap})
+//       : super(key: key);
 
-  final ScanResult result;
-  final VoidCallback? onTap;
+//   final ScanResult result;
+//   final VoidCallback? onTap;
 
-  Widget _buildTitle(BuildContext context) {
-    if (result.device.name.isNotEmpty) {
-      return Column(
-        mainAxisAlignment: MainAxisAlignment.start,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Text(
-            result.device.name,
-            overflow: TextOverflow.ellipsis,
-          ),
-          Text(
-            result.device.id.toString(),
-            style: Theme.of(context).textTheme.caption,
-          )
-        ],
-      );
-    } else {
-      return Text(result.device.id.toString());
-    }
-  }
+//   Widget _buildTitle(BuildContext context) {
+//     if (result.device.name.isNotEmpty) {
+//       return Column(
+//         mainAxisAlignment: MainAxisAlignment.start,
+//         crossAxisAlignment: CrossAxisAlignment.start,
+//         children: <Widget>[
+//           Text(
+//             result.device.name,
+//             overflow: TextOverflow.ellipsis,
+//           ),
+//           Text(
+//             result.device.id.toString(),
+//             style: Theme.of(context).textTheme.caption,
+//           )
+//         ],
+//       );
+//     } else {
+//       return Text(result.device.id.toString());
+//     }
+//   }
 
-  Widget _buildAdvRow(BuildContext context, String title, String value) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4.0),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Text(title, style: Theme.of(context).textTheme.caption),
-          const SizedBox(
-            width: 12.0,
-          ),
-          Expanded(
-            child: Text(
-              value,
-              style: Theme.of(context)
-                  .textTheme
-                  .caption
-                  ?.apply(color: Colors.black),
-              softWrap: true,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
+//   Widget _buildAdvRow(BuildContext context, String title, String value) {
+//     return Padding(
+//       padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4.0),
+//       child: Row(
+//         crossAxisAlignment: CrossAxisAlignment.start,
+//         children: <Widget>[
+//           Text(title, style: Theme.of(context).textTheme.caption),
+//           const SizedBox(
+//             width: 12.0,
+//           ),
+//           Expanded(
+//             child: Text(
+//               value,
+//               style: Theme.of(context)
+//                   .textTheme
+//                   .caption
+//                   ?.apply(color: Colors.black),
+//               softWrap: true,
+//             ),
+//           ),
+//         ],
+//       ),
+//     );
+//   }
 
-  String getNiceHexArray(List<int> bytes) {
-    return '[${bytes.map((i) => i.toRadixString(16).padLeft(2, '0')).join(', ')}]'
-        .toUpperCase();
-  }
+//   String getNiceHexArray(List<int> bytes) {
+//     return '[${bytes.map((i) => i.toRadixString(16).padLeft(2, '0')).join(', ')}]'
+//         .toUpperCase();
+//   }
 
-  String getNiceManufacturerData(Map<int, List<int>> data) {
-    if (data.isEmpty) {
-      return 'N/A';
-    }
-    List<String> res = [];
-    data.forEach((id, bytes) {
-      res.add(
-          '${id.toRadixString(16).toUpperCase()}: ${getNiceHexArray(bytes)}');
-    });
-    return res.join(', ');
-  }
+//   String getNiceManufacturerData(Map<int, List<int>> data) {
+//     if (data.isEmpty) {
+//       return 'N/A';
+//     }
+//     List<String> res = [];
+//     data.forEach((id, bytes) {
+//       res.add(
+//           '${id.toRadixString(16).toUpperCase()}: ${getNiceHexArray(bytes)}');
+//     });
+//     return res.join(', ');
+//   }
 
-  String getNiceServiceData(Map<String, List<int>> data) {
-    if (data.isEmpty) {
-      return 'N/A';
-    }
-    List<String> res = [];
-    data.forEach((id, bytes) {
-      res.add('${id.toUpperCase()}: ${getNiceHexArray(bytes)}');
-    });
-    return res.join(', ');
-  }
+//   String getNiceServiceData(Map<String, List<int>> data) {
+//     if (data.isEmpty) {
+//       return 'N/A';
+//     }
+//     List<String> res = [];
+//     data.forEach((id, bytes) {
+//       res.add('${id.toUpperCase()}: ${getNiceHexArray(bytes)}');
+//     });
+//     return res.join(', ');
+//   }
 
-  @override
-  Widget build(BuildContext context) {
-    return ExpansionTile(
-      title: _buildTitle(context),
-      leading: Text(result.rssi.toString()),
-      trailing: ElevatedButton(
-        child: const Text('CONNECT'),
-        style: ElevatedButton.styleFrom(
-          primary: Colors.black,
-          onPrimary: Colors.white,
-        ),
-        onPressed: (result.advertisementData.connectable) ? onTap : null,
-      ),
-      children: <Widget>[
-        _buildAdvRow(
-            context, 'Complete Local Name', result.advertisementData.localName),
-        _buildAdvRow(context, 'Tx Power Level',
-            '${result.advertisementData.txPowerLevel ?? 'N/A'}'),
-        _buildAdvRow(context, 'Manufacturer Data',
-            getNiceManufacturerData(result.advertisementData.manufacturerData)),
-        _buildAdvRow(
-            context,
-            'Service UUIDs',
-            (result.advertisementData.serviceUuids.isNotEmpty)
-                ? result.advertisementData.serviceUuids.join(', ').toUpperCase()
-                : 'N/A'),
-        _buildAdvRow(context, 'Service Data',
-            getNiceServiceData(result.advertisementData.serviceData)),
-      ],
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return ExpansionTile(
+//       title: _buildTitle(context),
+//       leading: Text(result.rssi.toString()),
+//       trailing: ElevatedButton(
+//         child: const Text('CONNECT'),
+//         style: ElevatedButton.styleFrom(
+//           primary: Colors.black,
+//           onPrimary: Colors.white,
+//         ),
+//         onPressed: (result.advertisementData.connectable) ? onTap : null,
+//       ),
+//       children: <Widget>[
+//         _buildAdvRow(
+//             context, 'Complete Local Name', result.advertisementData.localName),
+//         _buildAdvRow(context, 'Tx Power Level',
+//             '${result.advertisementData.txPowerLevel ?? 'N/A'}'),
+//         _buildAdvRow(context, 'Manufacturer Data',
+//             getNiceManufacturerData(result.advertisementData.manufacturerData)),
+//         _buildAdvRow(
+//             context,
+//             'Service UUIDs',
+//             (result.advertisementData.serviceUuids.isNotEmpty)
+//                 ? result.advertisementData.serviceUuids.join(', ').toUpperCase()
+//                 : 'N/A'),
+//         _buildAdvRow(context, 'Service Data',
+//             getNiceServiceData(result.advertisementData.serviceData)),
+//       ],
+//     );
+//   }
+// }
 
-class ServiceTile extends StatelessWidget {
-  final BluetoothService service;
-  final List<CharacteristicTile> characteristicTiles;
+// class ServiceTile extends StatelessWidget {
+//   final BluetoothService service;
+//   final List<CharacteristicTile> characteristicTiles;
 
-  const ServiceTile(
-      {Key? key, required this.service, required this.characteristicTiles})
-      : super(key: key);
+//   const ServiceTile(
+//       {Key? key, required this.service, required this.characteristicTiles})
+//       : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    if (characteristicTiles.isNotEmpty) {
-      return ExpansionTile(
-        title: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            const Text('Service'),
-            Text('0x${service.uuid.toString().toUpperCase().substring(4, 8)}',
-                style: Theme.of(context).textTheme.bodyText1?.copyWith(
-                    color: Theme.of(context).textTheme.caption?.color))
-          ],
-        ),
-        children: characteristicTiles,
-      );
-    } else {
-      return ListTile(
-        title: const Text('Service'),
-        subtitle:
-            Text('0x${service.uuid.toString().toUpperCase().substring(4, 8)}'),
-      );
-    }
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     if (characteristicTiles.isNotEmpty) {
+//       return ExpansionTile(
+//         title: Column(
+//           mainAxisAlignment: MainAxisAlignment.center,
+//           crossAxisAlignment: CrossAxisAlignment.start,
+//           children: <Widget>[
+//             const Text('Service'),
+//             Text('0x${service.uuid.toString().toUpperCase().substring(4, 8)}',
+//                 style: Theme.of(context).textTheme.bodyText1?.copyWith(
+//                     color: Theme.of(context).textTheme.caption?.color))
+//           ],
+//         ),
+//         children: characteristicTiles,
+//       );
+//     } else {
+//       return ListTile(
+//         title: const Text('Service'),
+//         subtitle:
+//             Text('0x${service.uuid.toString().toUpperCase().substring(4, 8)}'),
+//       );
+//     }
+//   }
+// }
 
-class CharacteristicTile extends StatelessWidget {
-  final BluetoothCharacteristic characteristic;
-  final List<DescriptorTile> descriptorTiles;
-  final VoidCallback? onReadPressed;
-  final VoidCallback? onWritePressed;
-  final VoidCallback? onNotificationPressed;
+// class CharacteristicTile extends StatelessWidget {
+//   final BluetoothCharacteristic characteristic;
+//   final List<DescriptorTile> descriptorTiles;
+//   final VoidCallback? onReadPressed;
+//   final VoidCallback? onWritePressed;
+//   final VoidCallback? onNotificationPressed;
 
-  const CharacteristicTile(
-      {Key? key,
-      required this.characteristic,
-      required this.descriptorTiles,
-      this.onReadPressed,
-      this.onWritePressed,
-      this.onNotificationPressed})
-      : super(key: key);
+//   const CharacteristicTile(
+//       {Key? key,
+//       required this.characteristic,
+//       required this.descriptorTiles,
+//       this.onReadPressed,
+//       this.onWritePressed,
+//       this.onNotificationPressed})
+//       : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<List<int>>(
-      stream: characteristic.value,
-      initialData: characteristic.lastValue,
-      builder: (c, snapshot) {
-        final value = snapshot.data;
-        return ExpansionTile(
-          title: ListTile(
-            title: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: <Widget>[
-                const Text('Characteristic'),
-                Text(
-                    '0x${characteristic.uuid.toString().toUpperCase().substring(4, 8)}',
-                    style: Theme.of(context).textTheme.bodyText1?.copyWith(
-                        color: Theme.of(context).textTheme.caption?.color))
-              ],
-            ),
-            subtitle: Text(value.toString()),
-            contentPadding: const EdgeInsets.all(0.0),
-          ),
-          trailing: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: <Widget>[
-              IconButton(
-                icon: Icon(
-                  Icons.file_download,
-                  color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
-                ),
-                onPressed: onReadPressed,
-              ),
-              IconButton(
-                icon: Icon(Icons.file_upload,
-                    color: Theme.of(context).iconTheme.color?.withOpacity(0.5)),
-                onPressed: onWritePressed,
-              ),
-              IconButton(
-                icon: Icon(
-                    characteristic.isNotifying
-                        ? Icons.sync_disabled
-                        : Icons.sync,
-                    color: Theme.of(context).iconTheme.color?.withOpacity(0.5)),
-                onPressed: onNotificationPressed,
-              )
-            ],
-          ),
-          children: descriptorTiles,
-        );
-      },
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return StreamBuilder<List<int>>(
+//       stream: characteristic.value,
+//       initialData: characteristic.lastValue,
+//       builder: (c, snapshot) {
+//         final value = snapshot.data;
+//         return ExpansionTile(
+//           title: ListTile(
+//             title: Column(
+//               mainAxisAlignment: MainAxisAlignment.center,
+//               crossAxisAlignment: CrossAxisAlignment.start,
+//               children: <Widget>[
+//                 const Text('Characteristic'),
+//                 Text(
+//                     '0x${characteristic.uuid.toString().toUpperCase().substring(4, 8)}',
+//                     style: Theme.of(context).textTheme.bodyText1?.copyWith(
+//                         color: Theme.of(context).textTheme.caption?.color))
+//               ],
+//             ),
+//             subtitle: Text(value.toString()),
+//             contentPadding: const EdgeInsets.all(0.0),
+//           ),
+//           trailing: Row(
+//             mainAxisSize: MainAxisSize.min,
+//             children: <Widget>[
+//               IconButton(
+//                 icon: Icon(
+//                   Icons.file_download,
+//                   color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
+//                 ),
+//                 onPressed: onReadPressed,
+//               ),
+//               IconButton(
+//                 icon: Icon(Icons.file_upload,
+//                     color: Theme.of(context).iconTheme.color?.withOpacity(0.5)),
+//                 onPressed: onWritePressed,
+//               ),
+//               IconButton(
+//                 icon: Icon(
+//                     characteristic.isNotifying
+//                         ? Icons.sync_disabled
+//                         : Icons.sync,
+//                     color: Theme.of(context).iconTheme.color?.withOpacity(0.5)),
+//                 onPressed: onNotificationPressed,
+//               )
+//             ],
+//           ),
+//           children: descriptorTiles,
+//         );
+//       },
+//     );
+//   }
+// }
 
-class DescriptorTile extends StatelessWidget {
-  final BluetoothDescriptor descriptor;
-  final VoidCallback? onReadPressed;
-  final VoidCallback? onWritePressed;
+// class DescriptorTile extends StatelessWidget {
+//   final BluetoothDescriptor descriptor;
+//   final VoidCallback? onReadPressed;
+//   final VoidCallback? onWritePressed;
 
-  const DescriptorTile(
-      {Key? key,
-      required this.descriptor,
-      this.onReadPressed,
-      this.onWritePressed})
-      : super(key: key);
+//   const DescriptorTile(
+//       {Key? key,
+//       required this.descriptor,
+//       this.onReadPressed,
+//       this.onWritePressed})
+//       : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      title: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          const Text('Descriptor'),
-          Text('0x${descriptor.uuid.toString().toUpperCase().substring(4, 8)}',
-              style: Theme.of(context)
-                  .textTheme
-                  .bodyText1
-                  ?.copyWith(color: Theme.of(context).textTheme.caption?.color))
-        ],
-      ),
-      subtitle: StreamBuilder<List<int>>(
-        stream: descriptor.value,
-        initialData: descriptor.lastValue,
-        builder: (c, snapshot) => Text(snapshot.data.toString()),
-      ),
-      trailing: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          IconButton(
-            icon: Icon(
-              Icons.file_download,
-              color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
-            ),
-            onPressed: onReadPressed,
-          ),
-          IconButton(
-            icon: Icon(
-              Icons.file_upload,
-              color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
-            ),
-            onPressed: onWritePressed,
-          )
-        ],
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return ListTile(
+//       title: Column(
+//         mainAxisAlignment: MainAxisAlignment.center,
+//         crossAxisAlignment: CrossAxisAlignment.start,
+//         children: <Widget>[
+//           const Text('Descriptor'),
+//           Text('0x${descriptor.uuid.toString().toUpperCase().substring(4, 8)}',
+//               style: Theme.of(context)
+//                   .textTheme
+//                   .bodyText1
+//                   ?.copyWith(color: Theme.of(context).textTheme.caption?.color))
+//         ],
+//       ),
+//       subtitle: StreamBuilder<List<int>>(
+//         stream: descriptor.value,
+//         initialData: descriptor.lastValue,
+//         builder: (c, snapshot) => Text(snapshot.data.toString()),
+//       ),
+//       trailing: Row(
+//         mainAxisSize: MainAxisSize.min,
+//         children: <Widget>[
+//           IconButton(
+//             icon: Icon(
+//               Icons.file_download,
+//               color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
+//             ),
+//             onPressed: onReadPressed,
+//           ),
+//           IconButton(
+//             icon: Icon(
+//               Icons.file_upload,
+//               color: Theme.of(context).iconTheme.color?.withOpacity(0.5),
+//             ),
+//             onPressed: onWritePressed,
+//           )
+//         ],
+//       ),
+//     );
+//   }
+// }
 
-class AdapterStateTile extends StatelessWidget {
-  const AdapterStateTile({Key? key, required this.state}) : super(key: key);
+// class AdapterStateTile extends StatelessWidget {
+//   const AdapterStateTile({Key? key, required this.state}) : super(key: key);
 
-  final BluetoothState state;
+//   final BluetoothState state;
 
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.redAccent,
-      child: ListTile(
-        title: Text(
-          'Bluetooth adapter is ${state.toString().substring(15)}',
-          style: Theme.of(context).primaryTextTheme.subtitle2,
-        ),
-        trailing: Icon(
-          Icons.error,
-          color: Theme.of(context).primaryTextTheme.subtitle2?.color,
-        ),
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return Container(
+//       color: Colors.redAccent,
+//       child: ListTile(
+//         title: Text(
+//           'Bluetooth adapter is ${state.toString().substring(15)}',
+//           style: Theme.of(context).primaryTextTheme.subtitle2,
+//         ),
+//         trailing: Icon(
+//           Icons.error,
+//           color: Theme.of(context).primaryTextTheme.subtitle2?.color,
+//         ),
+//       ),
+//     );
+//   }
+// }
