@@ -1,16 +1,18 @@
 import 'dart:async';
+import 'dart:convert';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
-import 'package:water_level_flutter/app/device_managment_page/data/device_repository.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:water_level_flutter/app/device_managment_page/presentation/device_managment_controller.dart';
 import 'package:water_level_flutter/app/homepage/application/devices_service.dart';
 import 'package:water_level_flutter/app/util/async_value_ui.dart';
 import 'package:water_level_flutter/routing/app_router.dart';
 
-// import 'package:flutter_zxing/flutter_zxing.dart';
+import 'package:flutter_zxing/flutter_zxing.dart';
 
 // import './domain/Device.dart';
 
@@ -37,10 +39,10 @@ class _DevicesManagmentPageState extends ConsumerState<DevicesManagmentPage> {
         children: [
           Container(
             decoration: BoxDecoration(
-              color: Colors.white,
+              color: Theme.of(context).primaryColor,
               boxShadow: [
                 BoxShadow(
-                  color: Colors.grey.withOpacity(0.5),
+                  color: Theme.of(context).shadowColor.withOpacity(0.5),
                   spreadRadius: 0,
                   blurRadius: 7,
                   offset: const Offset(0, 0.75), // changes position of shadow
@@ -54,18 +56,28 @@ class _DevicesManagmentPageState extends ConsumerState<DevicesManagmentPage> {
                 children: [
                   Row(),
                   Row(
-                    children: const [
-                      Icon(Icons.device_hub, size: 35),
+                    children: [
+                      Icon(
+                        Icons.device_hub,
+                        size: Theme.of(context).textTheme.titleMedium!.fontSize,
+                        // size: 35.sp,
+                      ),
                       Text(
                         'Devices',
-                        style: TextStyle(fontSize: 35),
+                        style: Theme.of(context).textTheme.titleMedium,
+                        // style: TextStyle(fontSize: 35.sp),
                       ),
                     ],
                   ),
                   Padding(
-                    padding: const EdgeInsets.only(right: 20.0),
+                    padding: EdgeInsets.only(right: 20.0.sp),
                     child: InkWell(
-                      child: const Icon(Icons.add, size: 40),
+                      child: Icon(
+                        Icons.add,
+                        size: Theme.of(context).textTheme.titleMedium!.fontSize,
+
+                        // size: 40.sp,
+                      ),
                       onTap: () {
                         showDialog(
                             context: context,
@@ -129,159 +141,54 @@ class _DevicesManagmentPageState extends ConsumerState<DevicesManagmentPage> {
           // Container(
           //     child: CircularProgressIndicator(),
           //   )
-          Column(
-            children: state.when(
-                data: (data) {
-                  return data
-                      .map(
-                        (e) => ListTile(
-                          leading: const Icon(Icons.water_drop_outlined),
-                          title: Text(e?.serialNumber ?? 'No Name'),
-                          trailing: InkWell(
-                            child: const Icon(Icons.edit),
-                            onTap: () {
-                              // DevicePage.show(context, e);
-                              context.goNamed(AppRoute.editDevice.name,
-                                  extra: e);
-                            },
+          Padding(
+            padding: EdgeInsets.all(8.0.sp),
+            child: Column(
+              children: state.when(
+                  data: (data) {
+                    return data
+                        .map(
+                          (e) => ListTile(
+                            leading: Icon(Icons.water_drop_outlined),
+                            title: SingleChildScrollView(
+                              scrollDirection: Axis.horizontal,
+                              child: Text(
+                                e?.serialNumber ?? 'No Name',
+                                style: Theme.of(context).textTheme.bodyMedium,
+                                // style: TextStyle(fontSize: 15.sp),
+                              ),
+                            ),
+                            trailing: InkWell(
+                              child: const Icon(
+                                Icons.edit,
+                                // size: Theme.of(context)
+                                //     .textTheme
+                                //     .bodyMedium!
+                                //     .fontSize,
+                              ),
+                              onTap: () {
+                                // DevicePage.show(context, e);
+                                context.goNamed(AppRoute.editDevice.name,
+                                    extra: e);
+                              },
+                            ),
                           ),
-                        ),
-                      )
-                      .toList();
-                },
-                error: ((error, stackTrace) {
-                  return [Text('Error $error')];
-                }),
-                loading: (() => [const CircularProgressIndicator()])),
+                        )
+                        .toList();
+                  },
+                  error: ((error, stackTrace) {
+                    return [Text('Error $error')];
+                  }),
+                  loading: (() => [const CircularProgressIndicator()])),
+            ),
           )
         ],
       ),
     );
   }
-
-  String _qrCodeResult = '';
-
-  // Future<void> _scanQRCode() async {
-  //   try {
-  //     await Navigator.push(
-  //       context,
-  //       MaterialPageRoute(
-  //         builder: (context) => Scaffold(body: ReaderWidget(
-  //           onScan: (result) {
-  //             _qrCodeResult = utf8.decode(base64.decode(result.text!));
-  //             Navigator.of(context).pop();
-  //           },
-  //         )),
-  //       ),
-  //     );
-  //     // });
-  //   } catch (e) {
-  //     print(e.toString());
-  //   }
-  // }
-  //
-  // Future<void> _dialogBuilder(BuildContext context) {
-  //   return showDialog<void>(
-  //       context: context,
-  //       builder: (BuildContext context) {
-  //         return Consumer(
-  //           builder: (context, ref, child) {
-  //             return AlertDialog(
-  //                 title: const Text('Claim a Device'),
-  //                 content: Container(
-  //                   height: 250,
-  //                   width: 250,
-  //                   child: ReaderWidget(
-  //                     onScan: (result) async {
-  //                       // Do something with the result
-  //
-  //                       final serialNumber =
-  //                           utf8.decode(base64.decode(result.text!));
-  //                       // Navigator.of(context).pop();
-  //                       print(serialNumber);
-  //                       final d = await ref
-  //                           .read(graphqlServices)
-  //                           .claimDevice(serialNumber);
-  //
-  //                       d.when(
-  //                           data: (data) {
-  //                             print(data.message);
-  //                             ref
-  //                                 .read(deviceManagementViewModelProvider
-  //                                     .notifier)
-  //                                 .getDevices();
-  //                             // Navigator.of(context).pop();
-  //                           },
-  //                           error: (erro, stackTrace) {
-  //                             List err = erro as List<GraphQLResponseError>;
-  //                             ref.read(errorPro.notifier).state =
-  //                                 err.first.message;
-  //                             ScaffoldMessenger.of(context).showSnackBar(
-  //                                 SnackBar(
-  //                                     content:
-  //                                         Text(err.first.message.toString())));
-  //                           },
-  //                           loading: () => const CircularProgressIndicator());
-  //
-  //                       // Navigator.of(context).pop();
-  //                     },
-  //                   ),
-  //                 )
-  //                 // TextFormField(
-  //                 //   controller: deviceSN,
-  //                 //   decoration: InputDecoration(
-  //                 //     label: Text('Device Serial Number'),
-  //                 //     errorText: ref.watch(errorPro),
-  //                 //   ),
-  //                 // ),
-  //                 // actions: <Widget>[
-  //                 //   TextButton(
-  //                 //     style: TextButton.styleFrom(
-  //                 //       textStyle: Theme.of(context).textTheme.labelLarge,
-  //                 //     ),
-  //                 //     child: const Text('Cancel'),
-  //                 //     onPressed: () {
-  //                 //       Navigator.of(context).pop();
-  //                 //     },
-  //                 //   ),
-  //                 //   TextButton(
-  //                 //     style: TextButton.styleFrom(
-  //                 //       textStyle: Theme.of(context).textTheme.labelLarge,
-  //                 //     ),
-  //                 //     child: const Text('Add'),
-  //                 //     onPressed: () async {
-  //                 //       final d = await ref
-  //                 //           .read(graphqlServices)
-  //                 //           .claimDevice(deviceSN.text);
-  //
-  //                 //       d.when(
-  //                 //           data: (data) {
-  //                 //             print(data.message);
-  //                 //             ref
-  //                 //                 .read(
-  //                 //                     deviceManagementViewModelProvider.notifier)
-  //                 //                 .getDevices();
-  //                 //             Navigator.of(context).pop();
-  //                 //           },
-  //                 //           error: (erro, stackTrace) {
-  //                 //             List err = erro as List<GraphQLResponseError>;
-  //                 //             ref.read(errorPro.notifier).state =
-  //                 //                 err.first.message;
-  //                 //             // ScaffoldMessenger.of(context).showSnackBar(
-  //                 //             //     SnackBar(content: Text(error.toString())));
-  //                 //           },
-  //                 //           loading: () => const CircularProgressIndicator());
-  //                 //     },
-  //                 //   ),
-  //                 // ],
-  //                 );
-  //           },
-  //         );
-  //       });
-  // }
 }
 
-class AddDeviceDialog extends ConsumerWidget {
+class AddDeviceDialog extends HookConsumerWidget {
   const AddDeviceDialog({Key? key}) : super(key: key);
 
   @override
@@ -305,30 +212,79 @@ class AddDeviceDialog extends ConsumerWidget {
       }
     }
 
-    return Consumer(
+    final TabController tabController =
+        useTabController(initialLength: 2, initialIndex: 1);
+
+    return HookConsumer(
       builder: (context, ref, child) {
         final state = ref.watch(deviceManagmentControllerProvider);
         final error = ref.watch(errorPro);
+        final useQRCode = useState(false);
         return AlertDialog(
           // <-- SEE HERE
-          title: const Text('Add Device'),
+          title: Text('Add Device',
+              style: Theme.of(context).textTheme.titleMedium),
+
           content: SingleChildScrollView(
-            child: ListBody(
-              children: <Widget>[
-                const Text('Enter Device ID'),
-                state.isLoading
-                    ? const Center(child: CircularProgressIndicator())
-                    : TextField(
-                        controller: deviceSN,
-                        decoration: InputDecoration(
-                          // border: OutlineInputBorder(),
-                          labelText: 'Device ID',
-                          errorText: error.isEmpty ? null : error,
-                        ),
-                        onChanged: (value) => ref
-                            .read(errorPro.notifier)
-                            .update((state) => state = validate(value)),
-                      ),
+            child: Column(
+              children: [
+                TabBar(
+                  onTap: (value) {
+                    print(value);
+                    if (value == 0) {
+                      useQRCode.value = true;
+                    } else {
+                      useQRCode.value = false;
+                    }
+                  },
+                  tabs: const [
+                    Tab(
+                      text: 'Scan QR Code',
+                    ),
+                    Tab(
+                      text: 'Enter Device ID',
+                    ),
+                  ],
+                  controller: tabController,
+                ),
+                ListBody(
+                  children: useQRCode.value
+                      ? [
+                          Container(
+                            height: 250,
+                            width: 250,
+                            child: ReaderWidget(
+                              onScan: (result) async {
+                                final serialNumber =
+                                    utf8.decode(base64.decode(result.text!));
+                                print(serialNumber);
+                                deviceSN.text = serialNumber;
+                                ref.read(errorPro.notifier).update(
+                                    (state) => state = validate(serialNumber));
+                                tabController.animateTo(1);
+                                useQRCode.value = false;
+                              },
+                            ),
+                          )
+                        ]
+                      : [
+                          state.isLoading
+                              ? const Center(child: CircularProgressIndicator())
+                              : TextField(
+                                  controller: deviceSN,
+                                  decoration: InputDecoration(
+                                    labelText: 'Device ID',
+                                    labelStyle:
+                                        Theme.of(context).textTheme.bodyMedium,
+                                    errorText: error.isEmpty ? null : error,
+                                  ),
+                                  onChanged: (value) => ref
+                                      .read(errorPro.notifier)
+                                      .update(
+                                          (state) => state = validate(value)),
+                                ),
+                        ],
+                ),
               ],
             ),
           ),
@@ -357,88 +313,6 @@ class AddDeviceDialog extends ConsumerWidget {
     );
   }
 }
-
-//   Future<void> _showAlertDialog() async {
-//     final deviceSN = TextEditingController();
-//     // ref.invalidate(deviceManagmentControllerProvider);
-//     final errorPro = StateProvider<String>((ref) {
-//       return validate('');
-//     });
-
-//     Future<void> submit(BuildContext context) async {
-//       final sucess = await ref
-//           .read(deviceManagmentControllerProvider.notifier)
-//           .claimDevice(deviceSN.text);
-//       if (sucess && mounted) {
-//         ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-//           content: Text('Device Added Successfully'),
-//           backgroundColor: Colors.green,
-//         ));
-//         Navigator.of(context).pop();
-//       }
-//     }
-
-//     return showDialog<void>(
-//       context: context,
-//       // barrierDismissible: false, // user must tap button!
-//       builder: (BuildContext context) {
-//         return Consumer(
-//           builder: (context, ref, child) {
-//             final state = ref.watch(deviceManagmentControllerProvider);
-//             final error = ref.watch(errorPro);
-//             return AlertDialog(
-//               // <-- SEE HERE
-//               title: const Text('Add Device'),
-//               content: SingleChildScrollView(
-//                 child: ListBody(
-//                   children: <Widget>[
-//                     const Text('Enter Device ID'),
-//                     state.isLoading
-//                         ? const Center(child: CircularProgressIndicator())
-//                         : TextField(
-//                             controller: deviceSN,
-//                             decoration: InputDecoration(
-//                               // border: OutlineInputBorder(),
-//                               labelText: 'Device ID',
-//                               errorText: error.isEmpty ? null : error,
-//                             ),
-//                             onChanged: (value) => ref
-//                                 .read(errorPro.notifier)
-//                                 .update((state) => state = validate(value)),
-//                           ),
-//                   ],
-//                 ),
-//               ),
-//               actions: <Widget>[
-//                 TextButton(
-//                   onPressed: state.isLoading
-//                       ? null
-//                       : () {
-//                           Navigator.of(context).pop();
-//                         },
-//                   child: const Text('No'),
-//                 ),
-//                 TextButton(
-//                   onPressed: error.isEmpty
-//                       ? state.isLoading
-//                           ? null
-//                           : () {
-//                               submit(context);
-//                               // ref
-//                               //     .read(deviceManagmentControllerProvider
-//                               //         .notifier)
-//                               //     .claimDevice(deviceSN.text);
-//                             }
-//                       : null,
-//                   child: const Text('Yes'),
-//                 ),
-//               ],
-//             );
-//           },
-//         );
-//       },
-//     );
-//   }
 
 String validate(String value) {
   //check if value contain only numbers and letters and symbols and not empty
